# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: Dreamshelix India EdTech Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /courses:
    get:
      operationId: listCourses
      tags: [courses]
      summary: List all published courses
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of courses
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CoursesListResponse"
    post:
      operationId: createCourse
      tags: [courses]
      summary: Create a new course (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "201":
          description: Course created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
  /courses/{courseId}:
    get:
      operationId: getCourse
      tags: [courses]
      summary: Get course details
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Course details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourseDetail"
    put:
      operationId: updateCourse
      tags: [courses]
      summary: Update a course (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCourseRequest"
      responses:
        "200":
          description: Course updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Course"
    delete:
      operationId: deleteCourse
      tags: [courses]
      summary: Delete a course (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Course deleted
  /enrollments:
    get:
      operationId: listMyEnrollments
      tags: [enrollments]
      summary: Get my enrolled courses
      security:
        - bearerAuth: []
      responses:
        "200":
          description: My enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
    post:
      operationId: enrollCourse
      tags: [enrollments]
      summary: Enroll in a course
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnrollRequest"
      responses:
        "201":
          description: Enrolled successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{courseId}/progress:
    put:
      operationId: updateProgress
      tags: [enrollments]
      summary: Update course progress
      security:
        - bearerAuth: []
      parameters:
        - name: courseId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProgressUpdateRequest"
      responses:
        "200":
          description: Progress updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: Get notifications
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    put:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      security:
        - bearerAuth: []
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Course:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
        instructor:
          type: string
        rating:
          type: number
        totalStudents:
          type: integer
        isPublished:
          type: boolean
        tags:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - category
        - price
        - level
        - instructor
        - rating
        - totalStudents
        - isPublished
        - createdAt
    CourseDetail:
      allOf:
        - $ref: "#/components/schemas/Course"
        - type: object
          properties:
            curriculum:
              type: array
              items:
                $ref: "#/components/schemas/CourseSection"
            whatYouLearn:
              type: array
              items:
                type: string
            requirements:
              type: array
              items:
                type: string
            isEnrolled:
              type: boolean
          required:
            - curriculum
            - whatYouLearn
            - requirements
    CourseSection:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        lessons:
          type: array
          items:
            $ref: "#/components/schemas/Lesson"
      required:
        - id
        - title
        - lessons
    Lesson:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        duration:
          type: string
        videoUrl:
          type: string
        isPreview:
          type: boolean
      required:
        - id
        - title
        - duration
        - isPreview
    CreateCourseRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        category:
          type: string
        price:
          type: number
        originalPrice:
          type: number
        duration:
          type: string
        level:
          type: string
          enum: [beginner, intermediate, advanced]
        language:
          type: string
        instructor:
          type: string
        tags:
          type: array
          items:
            type: string
        whatYouLearn:
          type: array
          items:
            type: string
        requirements:
          type: array
          items:
            type: string
        isPublished:
          type: boolean
      required:
        - title
        - description
        - category
        - price
        - level
        - instructor
    CoursesListResponse:
      type: object
      properties:
        courses:
          type: array
          items:
            $ref: "#/components/schemas/Course"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required:
        - courses
        - total
        - page
        - totalPages
    Enrollment:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        courseId:
          type: integer
        course:
          $ref: "#/components/schemas/Course"
        progressPercent:
          type: integer
        completedLessons:
          type: integer
        totalLessons:
          type: integer
        isCompleted:
          type: boolean
        certificateUrl:
          type: string
        enrolledAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - courseId
        - progressPercent
        - completedLessons
        - totalLessons
        - isCompleted
        - enrolledAt
    EnrollRequest:
      type: object
      properties:
        courseId:
          type: integer
        paymentId:
          type: integer
      required:
        - courseId
    ProgressUpdateRequest:
      type: object
      properties:
        lessonId:
          type: integer
        progressPercent:
          type: integer
      required:
        - lessonId
        - progressPercent
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
          enum: [enrollment, payment, referral, certificate, withdrawal, system]
        isRead:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - title
        - message
        - type
        - isRead
        - createdAt
...