# lib/api-zod/src/generated/types/courseSection.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dreamshelix India EdTech Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Lesson } from "./lesson";

export interface CourseSection {
  id: number;
  title: string;
  lessons: Lesson[];
}