# lib/api-zod/src/generated/types/coursesListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Dreamshelix India EdTech Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { Course } from "./course";

export interface CoursesListResponse {
  courses: Course[];
  total: number;
  page: number;
  totalPages: number;
}